Rename Wren Academy file
