+++ app/data/session-data-defaults.js
@@ -4,7 +4,7 @@ const OrgHelper = require('../data/helpers/organisation')
 const { DateTime } = require('luxon')
 let applications = require('./applications.json')
 const users = require('./users.json')
-const relationships = require('./relationships-single-default.js')
+const relationships = require('./relationships-wren-academy.js')
 const user = require('./user')
 
 // get related training providers
