+++ scripts/generate-courses.js
@@ -5,7 +5,7 @@ const path = require('path')
 
 const OrgHelper = require('../app/data/helpers/organisation')
 
-const relationships = require('../app/data/relationships-single-default.js')
+const relationships = require('../app/data/relationships-wren-academy.js')
 const partners = relationships.map(relationship => relationship.org2)
 
 const generateCourse = require('../app/data/generators/course')
